[BUGFIX] Ensure correct properties for missing files
If a file was deleted on Canto site and the identifier is still known
within TYPO3 a default data set is currently returned. This patch ensures
that only properties requested within the second argument are returned.
Furthermore the `folder_identifiers` property needs to be any empty
array instead of an empty string.